Refactor rvi/message.rs into its own submodule
This simplifies writing unit tests and adding new messages.
